+++ keyboards/deng/djam/info.json
@@ -1,5 +1,5 @@
 {
-    "keyboard_name": "DJMax",
+    "keyboard_name": "DJam",
     "url": "",
     "maintainer": "myst729",
     "layouts": {
