+++ keyboards/deng/djam/readme.md
@@ -1,20 +1,20 @@
-# djmax
+# DJam
 
-![djmax](https://i.imgur.com/n95Y83h.jpeg)
+![DJam](https://i.imgur.com/n95Y83h.jpeg)
 
 A custom controller for the music game DJMax.
 
 * Keyboard Maintainer: [Leo Deng](https://github.com/myst729)
-* Hardware Supported: djmax controller with ATMEGA32U4
+* Hardware Supported: DJMax controller with ATMEGA32U4
 * Hardware Availability: Private Groupbuy
 
 Make example for this keyboard (after setting up your build environment):
 
-    make deng/djmax:default
+    make deng/djam:default
 
 Flashing example for this keyboard:
 
-    make deng/djmax:default:flash
+    make deng/djam:default:flash
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
